requirements.txt
asgiref==3.3.1
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.4
et-xmlfile==1.1.0
geographiclib==1.52
geopy==2.2.0
jsonfield==3.1.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
psycopg2==2.8.6
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
six==1.16.0
sqlparse==0.2.4
xlrd==2.0.1
